Refactor expand feature to less files
