--- client/components/Reviews/Text/StyledReviewText.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const StyledReviewText = styled.p`
-  display: ${(props) => props.display};
-  margin: 0px;
-`;
-
-export default StyledReviewText;
